ccid: strip shared object to reduce closure size
The ccid derivation provides a pcsclite plugin to access CCID devices.
This shared object is written to a non-standard path "$out/pcsc", which
causes it not to get stripped by default. The consequence is that the
result is an unstripped binary that still contains references to gcc and
glibc-dev in its debug symbols, which unnecessarily inflates the closure
size from ~80MB to ~400MB.

Explicitly add the directory to the stripped paths to fix this issue.