nixos/firewall: add optional filter for outgoing traffic
While the firewall module is able to filter incoming traffic as well as
forwarding, one notable omission is that it cannot filter outgoing
traffic at all. There are some use cases for this though, for example:

  - To ensure that machines that use a global VPN will only ever use
    the VPN to send data.

  - To ensure that servers won't be able to send data except for the
    specific services that they provide.

Introduce a new "networking.firewall.filterOutput" option that allows
administrators to configure an outgoing firewall. The options provided
by this module mirror the same options that we have for incoming
traffic:

  - "allowTCPPorts", "allowUDPPorts", "allowTCPPortRange" as well as
    "allowUDPPortRange" configure which outgoing TCP and UDP ports are
    accepted.

  - "interfaces" allows to configure which destination TCP and UDP ports
    are accepted on a specific interface.

  - "trustedInterfaces" allows to configure trusted interfaces on which
    outgoing data is always accepted. As with incoming traffic, "lo" is
    marked as trusted interface by default.

  - "extraRules" allows to configure specific rules that cannot be
    configured via any of the above options.

Most of the rules could in theory work with both iptables and nftables.
For now though, they are only implemented for nftables.